+++ CHANGELOG.rst
@@ -16,11 +16,13 @@ see TODO.md for list of TODO items
 
   * added create_table convenience function for creating tables.
 
+  * return_row_action (TABLE_RETURN_ROW) changed to return the whole row including the tag.
+
 * v0.3.0:
 
   * added GetInput class
 
-  * changed **kwargs to **options for all calls. Removed options from cleaners and validators
+  * changed kwargs to options for all calls. Removed options from cleaners and validators
 
   * changed Cleaner, Convertor and Validator to abstract base classes (ABCMeta) and methods to abtract methods
 
@@ -129,7 +131,7 @@ see TODO.md for list of TODO items
 
 
  * Made a major change to how errors are handled. Added error_callbacks, convertor format strings, and
-  validation convertor strings. This changed most of the code base and some of the examples.
+    validation convertor strings. This changed most of the code base and some of the examples.
 
  * Added print_error, log_error, and ignore_error error callback routines.